backup/requirements.txt
supervision
ffmpeg
opencv-python-headless
torch
numpy
fastapi
uvicorn
nest-asyncio
pyngrok
python-multipart
Pillow
requests

